requirements.txt
Django==1.11.1
Pillow==4.1.1
djangular==0.2.7
django-phonenumber-field==1.3.0
djangorestframework==3.6.2
django-bootstrap-form
